+++ src/App/Context.spec.ts
@@ -6,14 +6,17 @@ import {
 
 	isDefined,
 	hasConstructor,
+	hasProperty,
 	hasMethod,
 	extendsClass
 } from "./../test/JasmineExtender";
-import * as Utils from "./../Utils";
-import * as Pointer from "./../Pointer";
-import * as App from "./../App";
+
 import AbstractContext from "./../AbstractContext";
+import * as Agents from "./../Agents";
+import * as App from "./../App";
 import PersistedApp from "./../PersistedApp";
+import * as Pointer from "./../Pointer";
+import * as Utils from "./../Utils";
 
 import AppContext from "./Context";
 
@@ -56,6 +59,16 @@ describe( module( "Carbon/App/Context" ), ():void => {
 			"Carbon.AbstractContext"
 		), ():void => {
 			expect( appContext instanceof AbstractContext );
+		});
+
+		it( hasProperty(
+			INSTANCE,
+			"agents",
+			"Carbon.Agents.Class",
+			"Instance of Agents class for manage the agents inside of an application."
+		), ():void => {
+			expect( appContext.agents ).toBeDefined();
+			expect( appContext.agents instanceof Agents.Class ).toBe( true );
 		});
 
 		it( hasMethod(