+++ src/Carbon.spec.ts
@@ -19,6 +19,7 @@ import {
 } from "./test/JasmineExtender";
 import setting from "./settings";
 
+import * as Agent from "./Agent";
 import * as Agents from "./Agents";
 import * as APIDescription from "./APIDescription";
 import * as App from "./App";
@@ -103,6 +104,15 @@ describe( module( "Carbon" ), ():void => {
 			expect( Utils.isString( carbon.version ) ).toBe( true );
 
 			expect( carbon.version ).toMatch( /\d+\.\d+\.\d+.*/ );
+		});
+
+		it( reexports(
+			STATIC,
+			"Agent",
+			"Carbon/Agent"
+		), ():void => {
+			expect( Carbon.Agent ).toBeDefined();
+			expect( Carbon.Agent ).toBe( Agent );
 		});
 
 		it( reexports(